+++ kinto/tests/core/resource/test_events.py
@@ -93,6 +93,13 @@ class ResourceChangedTest(BaseEventTest, unittest.TestCase):
 
     subscribed = (ResourceChanged,)
 
+    def test_events_have_custom_representation(self):
+        self.app.post_json(self.collection_url, self.body,
+                           headers=self.headers, status=201)
+        self.assertEqual(repr(self.events[0]),
+                         "<ResourceChanged action=create "
+                         "uri=%s>" % self.collection_url)
+
     def test_post_sends_create_action(self):
         self.app.post_json(self.collection_url, self.body,
                            headers=self.headers, status=201)
@@ -350,6 +357,25 @@ class BatchEventsTest(BaseEventTest, unittest.TestCase):
             ]
         }
         self.app.post_json("/batch", body, headers=self.headers)
+        self.assertEqual(len(self.events), 2)
+
+    def test_one_event_is_sent_per_parent_id(self):
+        # /mushrooms is a UserResource (see testapp.views), which means
+        # that parent_id depends on the authenticated user.
+        body = {
+            "defaults": {
+                "path": '/mushrooms',
+                "method": "POST",
+                "body": self.body
+            },
+            "requests": [
+                {"headers": {"Authorization": "Basic bWF0OjE="}},
+                {"headers": {"Authorization": "Basic dG90bzp0dXR1"}},
+                {"headers": {"Authorization": "Basic bWF0OjE="}},
+            ]
+        }
+        self.app.post_json("/batch", body, headers=self.headers)
+        # Two different auth headers, thus two different parent_id:
         self.assertEqual(len(self.events), 2)
 
     def test_one_event_is_sent_per_action(self):